Create a continuous exposure fit plot

Produces a logistic fit plot with a facettable exposures/quantiles/distributions in ggplot2

Usage

ggcontinuousexpdist(
  data = effICGI,
  response = "response",
  endpoint = "Endpoint",
  model_type = c("linear", "loess", "none"),
  DOSE = "DOSE",
  color_fill = "DOSE",
  fit_by_color_fill = FALSE,
  exposure_metrics = c("AUC", "CMAX"),
  exposure_metric_split = c("median", "tertile", "quartile", "none"),
  exposure_metric_soc_value = -99,
  exposure_metric_plac_value = 0,
  exposure_distribution = c("distributions", "lineranges", "none"),
  exposure_distribution_percent = TRUE,
  exposure_distribution_percent_text_size = 5,
  dose_plac_value = "Placebo",
  xlab = "Exposure Values",
  ylab = "Response",
  points_alpha = 0.2,
  points_show = TRUE,
  mean_obs_byexptile = TRUE,
  mean_obs_byexptile_plac = TRUE,
  mean_obs_byexptile_group = "none",
  mean_text_size = 5,
  mean_obs_bydose = TRUE,
  mean_obs_bydose_plac = FALSE,
  N_text_show = TRUE,
  N_text_size = 5,
  N_text_ypos = c("with means", "top"),
  binlimits_show = TRUE,
  binlimits_text_size = 5,
  binlimits_ypos = -Inf,
  binlimits_color = "#B3B3B380",
  dist_position_scaler = 0.2,
  dist_offset = 0,
  dist_scale = 0.9,
  lineranges_ypos = -1,
  lineranges_dodge = 1,
  lineranges_doselabel = FALSE,
  proj_bydose = TRUE,
  yproj = TRUE,
  yproj_xpos = 0,
  yproj_dodge = 0.2,
  yaxis_position = c("left", "right"),
  facet_formula = NULL,
  theme_certara = TRUE,
  return_list = FALSE
)

Arguments

data

Data to use with multiple endpoints stacked into response (values), Endpoint(endpoint name)

response

name of the column holding the response values

endpoint

name of the column holding the name/key of the endpoint default to Endpoint

model_type

type of the trend fit one of "linear", "loess", "none"

DOSE

name of the column holding the DOSE values default to DOSE

color_fill

name of the column to be used for color/fill default to DOSE column

fit_by_color_fill

fit split by color? default FALSE

exposure_metrics

name(s) of the column(s) to be stacked into expname exptile and split into exposure_metric_split

exposure_metric_split

one of "median", "tertile", "quartile", "none"

exposure_metric_soc_value

special exposure code for standard of care default -99

exposure_metric_plac_value

special exposure code for placebo default 0

exposure_distribution

one of distributions, lineranges or none

exposure_distribution_percent

show percent of distribution between binlimits TRUE/FALSE

exposure_distribution_percent_text_size

distribution percentages text size default to 5

dose_plac_value

string identifying placebo in DOSE column

xlab

text to be used as x axis label

ylab

text to be used as y axis label

points_alpha

alpha transparency for points

points_show

show the observations TRUE/FALSE

mean_obs_byexptile

observed mean by exptile TRUE/FALSE

mean_obs_byexptile_plac

observed mean by exptile placebo TRUE/FALSE

mean_obs_byexptile_group

additional grouping for exptile probabilities default none

mean_text_size

mean text size default to 5

mean_obs_bydose

observed mean by dose TRUE/FALSE

mean_obs_bydose_plac

observed mean by placebo dose TRUE/FALSE

N_text_show

show the N by exptile TRUE/FALSE

N_text_size

N by exposure text size default to 5

N_text_ypos

y position for N two text elements the first for by exptile and the second for by dose/color options include with means top bottom

binlimits_show

show the binlimits vertical lines TRUE/FALSE

binlimits_text_size

binlimits text size default to 5

binlimits_ypos

binlimits y position default to -Inf

binlimits_color

binlimits text color default to alpha("gray70",0.5)

dist_position_scaler

space occupied by the distribution default to 0.2

dist_offset

offset where the distribution position starts default to 0

dist_scale

scaling parameter for ggridges default to 0.9

lineranges_ypos

where to put the lineranges -1

lineranges_dodge

lineranges vertical dodge value 1

lineranges_doselabel

TRUE/FALSE

proj_bydose

project the probabilities on logistic curve TRUE/FALSE

yproj

project the probabilities on y axis TRUE/FALSE

yproj_xpos

y projection x position 0

yproj_dodge

y projection dodge value 0.2

yaxis_position

where to put y axis "left" or "right"

facet_formula

facet formula to be use otherwise endpoint ~ expname

theme_certara

apply certara colors and format for strips and default colour/fill

return_list

What to return if True a list of the datasets and plot is returned instead of only the plot

Examples

# Example 1
library(ggplot2)
library(patchwork)
effICGI <- logistic_data |>
dplyr::filter(!is.na(ICGI7))|>
dplyr::filter(!is.na(AUC))
effICGI$DOSE <- factor(effICGI$DOSE,
                      levels=c("0", "600", "1200","1800","2400"),
                      labels=c("Placebo", "600 mg", "1200 mg","1800 mg","2400 mg"))
effICGI$STUDY <- factor(effICGI$STUDY)    
effICGI <- tidyr::gather(effICGI,Endpoint,response,ICGI7,BRLS)
a <- ggcontinuousexpdist(data = effICGI |> dplyr::filter(Endpoint =="ICGI7"),
                 response = "response",
                 endpoint = "Endpoint",
                 exposure_metrics = c("AUC"),
                 exposure_metric_split = c("quartile"),
                 exposure_metric_soc_value = -99,
                 exposure_metric_plac_value = 0,
                 dist_position_scaler = 1, dist_offset = -1 ,
                 yproj_xpos =  -20 ,
                 yproj_dodge = 20 ,
                 exposure_distribution ="distributions")
#> Joining with `by = join_by(loopvariable, DOSE, quant_10)`
#> Joining with `by = join_by(loopvariable, DOSE, quant_90)`
#> Joining with `by = join_by(loopvariable, DOSE, quant_25)`
#> Joining with `by = join_by(loopvariable, DOSE, quant_75)`
#> Joining with `by = join_by(loopvariable, DOSE, medexp)`
#>    loopvariable Endpoint expname    DOSE   DOSE2 minexp maxexp medexp   meanexp
#>          <char>   <char>  <fctr>  <fctr>  <fctr>  <num>  <num>  <num>     <num>
#> 1:    ICGI7_AUC    ICGI7     AUC Placebo Placebo    0.0    0.0    0.0   0.00000
#> 2:    ICGI7_AUC    ICGI7     AUC  600 mg  600 mg    9.7   81.7   50.8  49.91812
#> 3:    ICGI7_AUC    ICGI7     AUC 1200 mg 1200 mg   12.1  173.7   94.2  96.52731
#> 4:    ICGI7_AUC    ICGI7     AUC 1800 mg 1800 mg   67.2  237.5  146.2 148.46667
#> 5:    ICGI7_AUC    ICGI7     AUC 2400 mg 2400 mg  107.8  342.7  166.7 176.28919
#>    Nmiss  Ntot     mean         SE quant_25 quant_75 quant_90 quant_10   ymid10
#>    <int> <int>    <num>      <num>    <num>    <num>    <num>    <num>    <num>
#> 1:     0   244 2.811475 0.08186089    0.000    0.000     0.00     0.00 2.631228
#> 2:     0   149 2.040268 0.09227322   44.000   56.400    62.52    39.10 2.370075
#> 3:     0   238 1.899160 0.07574048   79.600  112.700   134.26    69.15 2.169368
#> 4:     0    36 1.972222 0.20887496  121.725  167.725   201.95   108.90 1.903873
#> 5:     0    37 1.540541 0.12002624  140.400  205.100   251.76   117.40 1.847101
#>      ylow10    yup10    ymid90    ylow90    yup90   ymid25   ylow25    yup25
#>       <num>    <num>     <num>     <num>    <num>    <num>    <num>    <num>
#> 1: 2.503202 2.759255 2.6312284 2.5032018 2.759255 2.631228 2.503202 2.759255
#> 2: 2.275602 2.464548 2.2136504 2.1247430 2.302558 2.337348 2.245164 2.429531
#> 3: 2.079450 2.259285 1.7344910 1.5900320 1.878950 2.099571 2.005813 2.193329
#> 4: 1.787626 2.020121 1.2823820 1.0471263 1.517638 1.818214 1.688390 1.948037
#> 5: 1.722037 1.972164 0.9496955 0.6422713 1.257120 1.693481 1.541478 1.845484
#>      ymid75   ylow75    yup75   ymid50   ylow50    yup50
#>       <num>    <num>    <num>    <num>    <num>    <num>
#> 1: 2.631228 2.503202 2.759255 2.631228 2.503202 2.759255
#> 2: 2.254527 2.165523 2.343530 2.291930 2.201974 2.381886
#> 3: 1.878493 1.758399 1.998586 2.002056 1.898915 2.105197
#> 4: 1.510975 1.323321 1.698628 1.654742 1.495428 1.814056
#> 5: 1.261343 1.021602 1.501084 1.517821 1.331552 1.704089
#> Error in fortify(data): `data` must be a <data.frame>, or an object coercible by `fortify()`, or
#> a valid <data.frame>-like object coercible by `as.data.frame()`.
#> Caused by error in `.prevalidate_data_frame_like_object()`:
#> ! `colnames(data)` must return a <character> of length `ncol(data)`.

b <- ggcontinuousexpdist(data = effICGI |> dplyr::filter(Endpoint =="BRLS"),
                 response = "response",
                 endpoint = "Endpoint",
                 exposure_metrics = c("AUC"),
                 exposure_metric_split = c("quartile"),
                 exposure_metric_soc_value = -99,
                 exposure_metric_plac_value = 0,
                 dist_position_scaler = 4.2, dist_offset = 5 ,
                 yproj_xpos =  -20 ,
                 yproj_dodge = 20 ,
                 exposure_distribution ="distributions",
                 return_list = FALSE)            
#> Joining with `by = join_by(loopvariable, DOSE, quant_10)`
#> Joining with `by = join_by(loopvariable, DOSE, quant_90)`
#> Joining with `by = join_by(loopvariable, DOSE, quant_25)`
#> Joining with `by = join_by(loopvariable, DOSE, quant_75)`
#> Joining with `by = join_by(loopvariable, DOSE, medexp)`
#>    loopvariable Endpoint expname    DOSE   DOSE2 minexp maxexp medexp   meanexp
#>          <char>   <char>  <fctr>  <fctr>  <fctr>  <num>  <num>  <num>     <num>
#> 1:     BRLS_AUC     BRLS     AUC Placebo Placebo    0.0    0.0    0.0   0.00000
#> 2:     BRLS_AUC     BRLS     AUC  600 mg  600 mg    9.7   81.7   50.8  49.91812
#> 3:     BRLS_AUC     BRLS     AUC 1200 mg 1200 mg   12.1  173.7   94.2  96.52731
#> 4:     BRLS_AUC     BRLS     AUC 1800 mg 1800 mg   67.2  237.5  146.2 148.46667
#> 5:     BRLS_AUC     BRLS     AUC 2400 mg 2400 mg  107.8  342.7  166.7 176.28919
#>    Nmiss  Ntot     mean        SE quant_25 quant_75 quant_90 quant_10   ymid10
#>    <int> <int>    <num>     <num>    <num>    <num>    <num>    <num>    <num>
#> 1:     0   244 23.04098 0.3117261    0.000    0.000     0.00     0.00 23.12536
#> 2:     0   149 23.23649 0.4109359   44.000   56.400    62.52    39.10 23.16012
#> 3:     0   238 23.10127 0.3293215   79.600  112.700   134.26    69.15 23.18683
#> 4:     0    36 23.97059 0.7100592  121.725  167.725   201.95   108.90 23.22217
#> 5:     0    37 23.62162 0.9759506  140.400  205.100   251.76   117.40 23.22972
#>      ylow10    yup10   ymid90   ylow90    yup90   ymid25   ylow25    yup25
#>       <num>    <num>    <num>    <num>    <num>    <num>    <num>    <num>
#> 1: 22.59082 23.65989 23.12536 22.59082 23.65989 23.12536 22.59082 23.65989
#> 2: 22.76578 23.55445 23.18094 22.80947 23.55240 23.16447 22.77965 23.54929
#> 3: 22.81098 23.56268 23.24471 22.63942 23.85001 23.19612 22.80396 23.58829
#> 4: 22.73529 23.70905 23.30489 22.31909 24.29069 23.23357 22.68968 23.77746
#> 5: 22.70582 23.75363 23.34917 22.06111 24.63723 23.25017 22.61324 23.88710
#>      ymid75   ylow75    yup75   ymid50   ylow50    yup50
#>       <num>    <num>    <num>    <num>    <num>    <num>
#> 1: 23.12536 22.59082 23.65989 23.12536 22.59082 23.65989
#> 2: 23.17550 22.80376 23.54723 23.17052 22.79490 23.54613
#> 3: 23.22555 22.72251 23.72858 23.20910 22.77734 23.64086
#> 4: 23.27446 22.48809 24.06083 23.25533 22.58774 23.92291
#> 5: 23.30769 22.30310 24.31227 23.27355 22.49299 24.05412
#> Error in fortify(data): `data` must be a <data.frame>, or an object coercible by `fortify()`, or
#> a valid <data.frame>-like object coercible by `as.data.frame()`.
#> Caused by error in `.prevalidate_data_frame_like_object()`:
#> ! `colnames(data)` must return a <character> of length `ncol(data)`.
(a / b)  +
plot_layout(guides = "collect") &
 theme(legend.position = "top")
#> Error: object 'a' not found

#Example 2
effICGI$SEX <- as.factor(effICGI$SEX)
ggcontinuousexpdist(data = effICGI |>
  dplyr::filter(Endpoint =="ICGI7"),
                 response = "response",
                 endpoint = "Endpoint",
                 color_fill = "SEX",
                 exposure_metrics = c("AUC"),
                 exposure_metric_split = c("quartile"),
                 exposure_metric_soc_value = -99,
                 exposure_metric_plac_value = 0,
                 dist_position_scaler = 1, dist_offset = -1 ,
                 yproj_xpos =  -20 ,
                 yproj_dodge = 20 ,
                 exposure_distribution ="lineranges")
#> Joining with `by = join_by(loopvariable, SEX, color_fill2, DOSE, quant_10)`
#> Joining with `by = join_by(loopvariable, SEX, color_fill2, DOSE, quant_90)`
#> Joining with `by = join_by(loopvariable, SEX, color_fill2, DOSE, quant_25)`
#> Joining with `by = join_by(loopvariable, SEX, color_fill2, DOSE, quant_75)`
#> Joining with `by = join_by(loopvariable, SEX, color_fill2, DOSE, medexp)`
#>     loopvariable Endpoint expname    SEX color_fill2    DOSE   DOSE2 minexp
#>           <char>   <char>  <fctr> <fctr>      <fctr>  <fctr>  <fctr>  <num>
#>  1:    ICGI7_AUC    ICGI7     AUC      1           1 Placebo Placebo    0.0
#>  2:    ICGI7_AUC    ICGI7     AUC      1           1  600 mg  600 mg   11.9
#>  3:    ICGI7_AUC    ICGI7     AUC      1           1 1200 mg 1200 mg   12.1
#>  4:    ICGI7_AUC    ICGI7     AUC      1           1 1800 mg 1800 mg   83.4
#>  5:    ICGI7_AUC    ICGI7     AUC      1           1 2400 mg 2400 mg  111.0
#>  6:    ICGI7_AUC    ICGI7     AUC      2           2 Placebo Placebo    0.0
#>  7:    ICGI7_AUC    ICGI7     AUC      2           2  600 mg  600 mg    9.7
#>  8:    ICGI7_AUC    ICGI7     AUC      2           2 1200 mg 1200 mg   13.9
#>  9:    ICGI7_AUC    ICGI7     AUC      2           2 1800 mg 1800 mg   67.2
#> 10:    ICGI7_AUC    ICGI7     AUC      2           2 2400 mg 2400 mg  107.8
#>     maxexp medexp   meanexp Nmiss  Ntot     mean         SE quant_25 quant_75
#>      <num>  <num>     <num> <int> <int>    <num>      <num>    <num>    <num>
#>  1:    0.0   0.00   0.00000     0    93 3.139785 0.12754322    0.000    0.000
#>  2:   60.1  44.70  43.27541     0    61 2.081967 0.14275793   41.100   48.800
#>  3:  153.7  81.30  82.68365     0   104 2.144231 0.12022684   71.275   93.325
#>  4:  209.8 138.60 137.40000     0    11 1.909091 0.31491833  109.000  160.500
#>  5:  342.7 142.50 156.86154     0    13 1.692308 0.17484848  131.200  159.200
#>  6:    0.0   0.00   0.00000     0   151 2.609272 0.10335370    0.000    0.000
#>  7:   81.7  54.50  54.52273     0    88 2.011364 0.12148326   48.600   60.525
#>  8:  173.7 105.10 107.27164     0   134 1.708955 0.09405194   92.825  123.600
#>  9:  237.5 154.00 153.33600     0    25 2.000000 0.27080128  132.100  167.400
#> 10:  275.3 180.35 186.81250     0    24 1.458333 0.15901835  159.600  218.500
#>     quant_90 quant_10   ymid10   ylow10    yup10    ymid90    ylow90    yup90
#>        <num>    <num>    <num>    <num>    <num>     <num>     <num>    <num>
#>  1:     0.00     0.00 2.631228 2.503202 2.759255 2.6312284 2.5032018 2.759255
#>  2:    51.90    33.30 2.408814 2.310950 2.506678 2.2845825 2.1948777 2.374287
#>  3:   106.51    60.87 2.224671 2.135835 2.313507 1.9198363 1.8059214 2.033751
#>  4:   184.20    91.90 2.017418 1.916026 2.118810 1.4009362 1.1906619 1.611211
#>  5:   171.06   115.76 1.858055 1.734751 1.981359 1.4886997 1.2965210 1.680878
#>  6:     0.00     0.00 2.631228 2.503202 2.759255 2.6312284 2.5032018 2.759255
#>  7:    63.77    44.38 2.334809 2.242781 2.426838 2.2053015 2.1162918 2.294311
#>  8:   140.50    80.37 2.094428 2.000285 2.188571 1.6928134 1.5406855 1.844941
#>  9:   202.24   115.94 1.856852 1.733356 1.980348 1.2804451 1.0447770 1.516113
#> 10:   259.08   121.97 1.816577 1.686479 1.946675 0.9008044 0.5826051 1.219004
#>       ymid25   ylow25    yup25   ymid75    ylow75    yup75   ymid50   ylow50
#>        <num>    <num>    <num>    <num>     <num>    <num>    <num>    <num>
#>  1: 2.631228 2.503202 2.759255 2.631228 2.5032018 2.759255 2.631228 2.503202
#>  2: 2.356717 2.263245 2.450189 2.305288 2.2147959 2.395780 2.332672 2.240770
#>  3: 2.155175 2.064693 2.245656 2.007900 1.9054354 2.110365 2.088217 1.993591
#>  4: 1.903205 1.786859 2.019552 1.559231 1.3812710 1.737191 1.705504 1.555734
#>  5: 1.754929 1.614145 1.895713 1.567914 1.3916804 1.744148 1.679455 1.524824
#>  6: 2.631228 2.503202 2.759255 2.631228 2.5032018 2.759255 2.631228 2.503202
#>  7: 2.306624 2.216073 2.397175 2.226975 2.1381455 2.315805 2.267217 2.177982
#>  8: 2.011240 1.909155 2.113324 1.805690 1.6737531 1.937628 1.929254 1.816681
#>  9: 1.748918 1.607059 1.890776 1.513145 1.3259313 1.700359 1.602645 1.433259
#> 10: 1.565242 1.388478 1.742007 1.171843 0.9128819 1.430804 1.426651 1.221719
#>        yup50
#>        <num>
#>  1: 2.759255
#>  2: 2.424574
#>  3: 2.182842
#>  4: 1.855273
#>  5: 1.834086
#>  6: 2.759255
#>  7: 2.356452
#>  8: 2.041826
#>  9: 1.772032
#> 10: 1.631583
#> Error in fortify(data): `data` must be a <data.frame>, or an object coercible by `fortify()`, or
#> a valid <data.frame>-like object coercible by `as.data.frame()`.
#> Caused by error in `.prevalidate_data_frame_like_object()`:
#> ! `colnames(data)` must return a <character> of length `ncol(data)`.
if (FALSE) { # \dontrun{
#Example 3
} # }