Create a fit vs exposure(s)plot

Produces a fit as per model_type plot with a facettable exposures/quantiles/distributions in ggplot2

Usage

ggresponseexpdist(
  data = effICGI,
  response = "response",
  endpoint = "Endpoint",
  model_type = c("loess", "linear", "logistic", "none"),
  DOSE = "DOSE",
  color_fill = "DOSE",
  fit_by_color_fill = FALSE,
  exposure_metrics = c("AUC", "CMAX"),
  exposure_metric_split = c("median", "tertile", "quartile", "none"),
  exposure_metric_soc_value = -99,
  exposure_metric_plac_value = 0,
  exposure_distribution = c("distributions", "lineranges", "boxplots", "none"),
  exposure_distribution_percent = TRUE,
  exposure_distribution_percent_text_size = 5,
  dose_plac_value = "Placebo",
  xlab = "Exposure Values",
  ylab = "Response",
  points_alpha = 0.2,
  points_show = TRUE,
  mean_obs_byexptile = TRUE,
  mean_obs_byexptile_plac = TRUE,
  mean_obs_byexptile_group = "none",
  mean_obs_bydose = TRUE,
  mean_obs_bydose_plac = FALSE,
  mean_obs_text_size = 5,
  N_byexptile_ypos = c("with means", "top", "bottom", "none"),
  N_bydose_ypos = c("with means", "top", "bottom", "none"),
  N_text_size = 5,
  N_text_sep = "\n",
  binlimits_show = TRUE,
  binlimits_text_size = 5,
  binlimits_ypos = 0,
  binlimits_color = "#B3B3B380",
  dist_position_scaler = 0.2,
  dist_offset = 0,
  dist_scale = 0.9,
  lineranges_ypos = -1,
  lineranges_dodge = 1,
  lineranges_doselabel = FALSE,
  proj_bydose = TRUE,
  yproj = TRUE,
  yproj_xpos = 0,
  yproj_dodge = 0.2,
  yaxis_position = c("left", "right"),
  facet_formula = NULL,
  theme_certara = TRUE,
  return_list = FALSE
)

Arguments

data

Data to use with multiple endpoints stacked into response (values), Endpoint(endpoint name)

response

name of the column holding the response values

endpoint

name of the column holding the name/key of the endpoint default to Endpoint

model_type

type of the trend fit one of "loess", "linear", "logistic", "none"

DOSE

name of the column holding the DOSE values default to DOSE

color_fill

name of the column to be used for color/fill default to DOSE column

fit_by_color_fill

fit split by color? default FALSE

exposure_metrics

name(s) of the column(s) to be stacked into expname exptile and split into exposure_metric_split

exposure_metric_split

one of "median", "tertile", "quartile", "none"

exposure_metric_soc_value

special exposure code for standard of care default -99

exposure_metric_plac_value

special exposure code for placebo default 0

exposure_distribution

one of distributions, lineranges, boxplots or none

exposure_distribution_percent

show percent of distribution between binlimits TRUE/FALSE

exposure_distribution_percent_text_size

distribution percentages text size default to 5

dose_plac_value

string identifying placebo in DOSE column

xlab

text to be used as x axis label

ylab

text to be used as y axis label

points_alpha

alpha transparency for points

points_show

show the observations TRUE/FALSE

mean_obs_byexptile

observed mean by exptile TRUE/FALSE

mean_obs_byexptile_plac

observed mean by exptile placebo TRUE/FALSE

mean_obs_byexptile_group

additional grouping for exptile means default none

mean_obs_bydose

observed mean by dose TRUE/FALSE

mean_obs_bydose_plac

observed mean by placebo dose TRUE/FALSE

mean_obs_text_size

mean text size default to 5

N_byexptile_ypos

y position for N responders/Ntotal for by exptile one of with means top bottom none

N_bydose_ypos

y position for N responders/Ntotal for by dose/color one of with means top bottom none

N_text_size

N responders (logistic) and or Ntotal text size default to 5

N_text_sep

character string to separate N from mean for logistic fit defaults to / otherwise \n

binlimits_show

show the binlimits vertical lines TRUE/FALSE

binlimits_text_size

binlimits text size default to 5

binlimits_ypos

binlimits y position default to -Inf

binlimits_color

binlimits text color default to alpha("gray70",0.5)

proj_bydose

project the predictions on the fit curve TRUE/FALSE

yproj

project the predictions on y axis TRUE/FALSE

yproj_xpos

y projection x position 0

yproj_dodge

y projection dodge value 0.2

yaxis_position

where to put y axis "left" or "right"

facet_formula

facet formula to be use otherwise endpoint ~ expname

theme_certara

apply certara colors and format for strips and default colour/fill

return_list

What to return if True a list of the datasets and plot is returned instead of only the plot

Examples

# Example 1
if (FALSE) { # \dontrun{
#' # Example 3                
} # }